+++ commandAPI.ts
@@ -1,0 +1,6 @@
+export interface BotCommand {
+
+    help(): string;
+    isCommand(command: string): boolean;
+    runCommand(args: string[], bot: any, msg: any): void;
+}